backend-python/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
openai==1.82.0
python-multipart==0.0.20
pypdf==5.6.0
python-dotenv==1.1.0
slowapi==0.1.9
